Moved dbf mass back into external subsystem
