+++ aviary/examples/external_subsystems/dbf_based_mass/airfoil_input.py
