--- outputted_phase_info.py
@@ -1,68 +1,0 @@
-phase_info = {
-    'pre_mission': {'include_takeoff': False, 'optimize_mass': True},
-    'climb_1': {
-        'subsystem_options': {'core_aerodynamics': {'method': 'computed'}},
-        'user_options': {
-            'num_segments': 3,
-            'order': 3,
-            'mach_optimize': False,
-            'mach_initial': (0.3, 'unitless'),
-            'mach_final': (0.5, 'unitless'),
-            'mach_bounds': ((0.27999999999999997, 0.52), 'unitless'),
-            'altitude_optimize': False,
-            'altitude_initial': (0.0, 'ft'),
-            'altitude_final': (39688.0, 'ft'),
-            'altitude_bounds': ((0.0, 40188.0), 'ft'),
-            'throttle_enforcement': 'path_constraint',
-            'time_initial_bounds': ((0.0, 0.0), 'min'),
-            'time_duration_bounds': ((50.5, 151.5), 'min'),
-            'distance_solve_segments': False,
-        },
-        'initial_guesses': {'time': ([0.0, 101.0], 'min')},
-    },
-    'climb_2': {
-        'subsystem_options': {'core_aerodynamics': {'method': 'computed'}},
-        'user_options': {
-            'num_segments': 3,
-            'order': 3,
-            'mach_optimize': False,
-            'mach_initial': (0.5, 'unitless'),
-            'mach_final': (0.5, 'unitless'),
-            'mach_bounds': ((0.48, 0.52), 'unitless'),
-            'altitude_optimize': False,
-            'altitude_initial': (39688.0, 'ft'),
-            'altitude_final': (40067.0, 'ft'),
-            'altitude_bounds': ((39188.0, 40567.0), 'ft'),
-            'throttle_enforcement': 'boundary_constraint',
-            'time_initial_bounds': ((50.5, 151.5), 'min'),
-            'time_duration_bounds': ((99.0, 297.0), 'min'),
-            'distance_solve_segments': False,
-        },
-        'initial_guesses': {'time': ([101.0, 198.0], 'min')},
-    },
-    'descent_1': {
-        'subsystem_options': {'core_aerodynamics': {'method': 'computed'}},
-        'user_options': {
-            'num_segments': 3,
-            'order': 3,
-            'mach_optimize': False,
-            'mach_initial': (0.5, 'unitless'),
-            'mach_final': (0.5, 'unitless'),
-            'mach_bounds': ((0.48, 0.52), 'unitless'),
-            'altitude_optimize': False,
-            'altitude_initial': (40067.0, 'ft'),
-            'altitude_final': (4432.0, 'ft'),
-            'altitude_bounds': ((3932.0, 40567.0), 'ft'),
-            'throttle_enforcement': 'path_constraint',
-            'time_initial_bounds': ((149.5, 448.5), 'min'),
-            'time_duration_bounds': ((46.5, 139.5), 'min'),
-            'distance_solve_segments': False,
-        },
-        'initial_guesses': {'time': ([299.0, 90.0], 'min')},
-        'post_mission': {
-            'include_landing': False,
-            'constrain_range': True,
-            'target_range': (2065.78, 'nmi'),
-        },
-    },
-}
